% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/court.R
\docType{data}
\name{court}
\alias{court}
\title{Polygon to draw the basketball court.}
\format{
A data frame with 6548 rows and 5 columns.
}
\usage{
data("court")
}
\description{
A `data.frame` containing the lines and circles used to draw
  a basketball court.
}
\examples{
data(court)
# Default plot
ggplot(court, aes(x = x, y = y, group = group)) + geom_polygon() + coord_fixed()
}
\references{
https://egallic.fr/en/drawing-a-basketball-court-with-r/
}
