% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcourt.R
\name{ggcourt}
\alias{ggcourt}
\title{Basketball Court}
\usage{
ggcourt(orientation = "wide", halfcourt = FALSE, ...)
}
\arguments{
\item{orientation}{orientation of the plot, can be "wide" (default) or "tall".}

\item{halfcourt}{boolean. If TRUE, only draws a half court.}
}
\description{
Draw a base basketball court.
}
\details{
This only plots a plain court without any labels or ticks.
}
\examples{
ggcourt(orientation = "wide")

}
\seealso{
\code{\link{court}}
}
