% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggshotintensity.R
\name{ggshotintensity}
\alias{ggshotintensity}
\title{A Heatmap-type Visualization of Shot Intensity}
\usage{
ggshotintensity(
  intensity,
  breaks = NULL,
  palette = NULL,
  x,
  y,
  value_col,
  color = "black",
  orientation = "wide",
  label_name = NULL
)
}
\arguments{
\item{intensity}{a `data.frame` with at least three columns. They should
be the (x,y) locations of the center of each rectangle in the grid, and the
corresponding (estimated) intensity values on the rectangle.}

\item{breaks}{if there are too many intensity values, for ease of
visualization, it is often better to break them into separate buckets. If
not specified, defaults to around 10.}

\item{palette}{optional color palette. If not specified, defaults to be based
on `Blues` in \code{RColorBrewer}}

\item{x}{column name for horizontal locations of rectangle centers}

\item{y}{column name for vertical locations of rectangle centers}

\item{value_col}{name of the column that contains the intensity values}

\item{color}{color for the boundary of contours, defaults to black.}

\item{orientation}{orientation of court, defaults to "wide"}

\item{label_name}{legend label in the final plot.}
}
\description{
Draw a heatmap on the court visualizing shot intensities.
}
\details{
Shot intensity plots help us identify the strengths and weaknesses
  of players. It is often a summary of the shotchart, either using raw
  counts, or model-based estimated values.
}
\examples{
# load data, take Draymond Green's LGCP estimated intensity as an example.
data(green_intensity)
head(green_intensity)

ggshotintensity(green_intensity, x = "x", y = "y", value_col = "mean")

}
\seealso{
\code{\link{court}}
}
