% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_intensity.R
\docType{data}
\name{green_intensity}
\alias{green_intensity}
\title{Shot data for selected players.}
\format{
a data.frame with columns being: the estimated intensity,
  (x, y) locations of rectangle centers over the court, Player name.
}
\usage{
data("green_intensity")
}
\description{
A `data.frame` containing the estimated shooting intensity of
  Draymond Green on the offensive half court.
}
\examples{
data(green_intensity)
ggshotintensity(green_intensity, x = "x", y = "y", value_col = "mean")

}
\references{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1002/sta4.324}
}
