% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shotdata.R
\docType{data}
\name{shotdata}
\alias{shotdata}
\title{Shot data for selected players.}
\format{
a data.frame with columns being: player name, indicator for
  made/missed shots, shot type, distance, x and y locations of shot, and
  string describing made/missed shot results
}
\usage{
data("shotdata")
}
\description{
A `data.frame` containing the shot attempts made by 191 players
  during the 2017--2018 NBA regular season (overtime periods not included).
}
\examples{
data(shotdata)
# Default plot

}
\references{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1002/sta4.324}
}
