% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggshotchart.R
\name{ggshotchart}
\alias{ggshotchart}
\title{Shot Chart}
\usage{
ggshotchart(shotdata, x, y, orientation = "wide", bytype = FALSE, ...)
}
\arguments{
\item{shotdata}{a data.frame containing the location of shots.}

\item{x, y}{x and y variables for drawing}

\item{color}{optional, color for shot locations}

\item{shape}{optional, shape for shot locations}
}
\description{
Draw a shot chart on court.
}
\details{
Shot charts helps provide insights when analyzing a player's strength
  and weakness.
}
\examples{


}
\seealso{

}
